Second homework
