+++ src/App.js
@@ -57,38 +57,42 @@
 // export default App;
 
 import * as React from 'react';
+import TodoList from './TodoList';
+// import AddTodoForm from './AddTodoForm'
 
 // Create an empty todoList array with at least 3 objects
-const todoList = [
-  {
-    id: 1,
-    title: 'Complete assignment for React class'
-  },
-  {
-    id: 2,
-    title: 'Submit project proposal'
-  },
-  {
-    id: 3,
-    title: 'Buy groceries'
-  }
-];
+// const todoList = [
+//   {
+//     id: 1,
+//     title: 'Complete assignment for React class'
+//   },
+//   {
+//     id: 2,
+//     title: 'Submit project proposal'
+//   },
+//   {
+//     id: 3,
+//     title: 'Buy groceries'
+//   }
+// ];
 
 // TodoList component that renders the list of todos
-function TodoList() {
-  return (
-    <ul>
-      {todoList.map(function(item) {
-        // For each item, return a <li> with the unique key and title
-        return (
-          <li key={item.id}>
-            {item.title}
-          </li>
-        );
-      })}
-    </ul>
-  );
-}
+// function TodoList() {
+//   return (
+//     // <ul>
+//     //   {todoList.map(function(item) {
+//     //     // For each item, return a <li> with the unique key and title
+//     //     return (
+//     //       <li key={item.id}>
+//     //         {item.title}
+//     //       </li>
+//     //     );
+//     //   })}
+//     // </ul>
+    
+//   );
+// }
+
 
 // App component where the TodoList will be displayed
 function App() {
@@ -96,6 +100,8 @@ function App() {
     <div>
       <h1>Todo List</h1>
       <TodoList />
+      {/* <AddTodoForm /> */}
+     
       <h4><i>Henry Osorio</i></h4>
     </div>
   );
